requirements.txt
APScheduler==3.6.3
attrs==22.2.0
black==23.1.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
flake8==3.9.2
flake8-broken-line==0.6.0
flake8-commas==2.1.0
flake8-comments==0.1.2
flake8-docstrings==1.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-print==5.0.0
flake8-return==1.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.6.1
mypy==1.0.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
pep8-naming==0.13.2
platformdirs==3.0.0
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pycodestyle==2.7.0
pydocstyle==6.3.0
pyflakes==2.3.1
PySocks==1.7.1
pytest==6.2.5
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
snowballstemmer==2.2.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
types-backports.ssl-match-hostname==3.7.4.3
types-cachetools==5.3.0.4
types-colorama==0.4.15.8
types-pytz==2022.7.1.2
types-requests==2.28.11.15
types-ujson==5.7.0.1
types-urllib3==1.26.25.8
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
